analyzer/requirements.txt
flask==3.0.3
kafka-python==2.0.2
redis==5.0.3
